Split tests into individual class files
